chore: get Kurt usage examples for `kurt` package README from example files
This commit splits the `model-agnostic` example sections into separate files
so that they can easily be pulled into the README for the `kurt` package
from there.

Each example uses a common `createKurt` function for setup, which was
split into a separate utility file of its own.
