+++ examples/basic/tsconfig.json
@@ -10,6 +10,6 @@
     "moduleResolution": "bundler",
     "target": "es2022"
   },
-  "include": ["src/*.ts"],
+  "include": ["src/**/*.ts"],
   "exclude": ["node_modules", "dist"]
 }
